total update

+++ ft_bzero.c
@@ -6,13 +6,15 @@
 /*   By: bkarimov <bkarimov@student.42prague.c      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/21 11:50:10 by bkarimov          #+#    #+#             */
-/*   Updated: 2025/05/21 13:15:07 by bkarimov         ###   ########.fr       */
+/*   Updated: 2025/06/16 10:54:21 by bkarimov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-void	ft_bzero(void *s, unsigned int n)
+#include "libft.h"
+
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*str;
-	int				c;
+	size_t			c;
 
 	str = (unsigned char *)s;
 	c = 0;
