+++ ft_memchr.c
@@ -6,13 +6,15 @@
 /*   By: bkarimov <bkarimov@student.42prague.c      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/26 14:17:44 by bkarimov          #+#    #+#             */
-/*   Updated: 2025/05/26 14:37:49 by bkarimov         ###   ########.fr       */
+/*   Updated: 2025/06/16 10:56:18 by bkarimov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-void	*ft_memchr(const void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char		*ptr;
-	int					count;
+	size_t				count;
 
 	ptr = (unsigned char *)s;
 	count = 0;
