+++ ft_memcmp.c
@@ -6,14 +6,16 @@
 /*   By: bkarimov <bkarimov@student.42prague.c      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/26 14:39:06 by bkarimov          #+#    #+#             */
-/*   Updated: 2025/05/26 14:51:33 by bkarimov         ###   ########.fr       */
+/*   Updated: 2025/06/16 10:57:02 by bkarimov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	unsigned char	*ptr1;
 	unsigned char	*ptr2;
-	unsigned int	i;
+	size_t			i;
 
 	ptr1 = (unsigned char *)s1;
 	ptr2 = (unsigned char *)s2;
